POSレジスタ機能を実装し、商品検索、カート管理、会計処理を追加
